cast to int and updated tests

+++ pitest-html-report/src/main/java/org/pitest/mutationtest/report/html/MutationTotals.java
@@ -49,13 +49,13 @@ public class MutationTotals {
   }
 
   public int getLineCoverage() {
-    return this.numberOfLines == 0 ? 100 : Math
+    return this.numberOfLines == 0 ? 100 : (int) Math
         .floor((100f * this.numberOfLinesCovered) / this.numberOfLines);
   }
 
   public int getMutationCoverage() {
     return this.numberOfMutations == 0 ? 100
-        : Math.floor((100f * this.numberOfMutationsDetected)
+        : (int) Math.floor((100f * this.numberOfMutationsDetected)
             / this.numberOfMutations);
   }
